requirements.txt
Django==5.2.6
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
pdf2image==1.17.0
Pillow==11.3.0
PyPDF2==3.0.1
pytesseract==0.3.13
python_docx==1.2.0
Requests==2.32.5
